Render updates for events
